add tag support for product settings and issue settings

+++ packages/frontend/src/scripts/components/widgets/TagView.tsx
@@ -3,11 +3,12 @@ import * as React from 'react'
 import { Tag } from 'productboard-common'
 
 import { TagManager } from '../../managers/tag'
+import { TagAssignmentManager } from '../../managers/tagAssignment'
 
 import DownIcon from '/src/images/down.png'
 import UpIcon from '/src/images/up.png'
 
-export const TagView = (props: { value: Tag[], productId: string, assignable: boolean }) => {
+export const TagView = (props: { tags: Tag[], productId: string, assignable: boolean, assignedTags?: Tag[], issueId?: string }) => {
 
     // PARAMS
 
@@ -16,13 +17,14 @@ export const TagView = (props: { value: Tag[], productId: string, assignable: bo
     // STATES
 
     // - Entities
-    const [tags, setTags] = React.useState<Tag[]>(props.value)
+    const [tags, setTags] = React.useState<Tag[]>(props.tags)
+    const [assignedTags, setAssignedTags] = React.useState<Tag[]>(props.assignedTags)
     // - Values
     const [tagName, setTagName] = React.useState<string>('new tag')
     const [tagColor, setTagColor] = React.useState<string>('blue')
     // - Interactions
     const [selectedTag, setSelectedTag] = React.useState<Tag>(null)
-    const [expanded, setExpanded] = React.useState<boolean>(true)
+    const [expanded, setExpanded] = React.useState<boolean>(false)
 
     function toggleDropdown() {
         setExpanded(expanded => !expanded)
@@ -62,17 +64,42 @@ export const TagView = (props: { value: Tag[], productId: string, assignable: bo
                         : tag
                 })
             })
+            setAssignedTags((prev) => {
+                return prev.map((tag) => {
+                    return tag.id === selectedTag.id
+                        ? { tag, ...selectedTag, ...updatedData }
+                        : tag
+                })
+            })
         }
     }
 
     async function deleteTag(event: React.FormEvent) {
         event.preventDefault()
         if (selectedTag) {
-            await TagManager.deleteTag(selectedTag.id)
-            setTags(tags.filter(other => other.id != selectedTag.id))
-            setSelectedTag(null)
-            setTagColor('blue')
-            setTagName('new tag')
+            if (confirm('Do you really want to delete this tag?')) {
+                await TagManager.deleteTag(selectedTag.id)
+                setTags(tags.filter(other => other.id != selectedTag.id))
+                setAssignedTags(assignedTags.filter(other => other.id != selectedTag.id))
+                setSelectedTag(null)
+                setTagColor('blue')
+                setTagName('new tag')
+            }
+        }
+    }
+    async function assignTag(event: React.FormEvent) {
+        event.preventDefault()
+        if (selectedTag) {
+            const assignment = await TagAssignmentManager.addTagAssignment({ tagId: selectedTag.id, issueId: props.issueId })
+            setAssignedTags((prev) => [...prev, tags.find(tag => assignment.tagId === tag.id)])
+        }
+    }
+    async function deassignTag(event: React.FormEvent) {
+        event.preventDefault()
+        if (selectedTag && assignedTags.includes(selectedTag)) {
+            const assignment = await TagAssignmentManager.findTagAssignments(productId).then(assignments => assignments.find(assignment => assignment.tagId === selectedTag.id))
+            await TagAssignmentManager.deleteTagAssignment(assignment.id)
+            setAssignedTags(assignedTags.filter(other => other.id != selectedTag.id))
         }
     }
 
@@ -81,7 +108,18 @@ export const TagView = (props: { value: Tag[], productId: string, assignable: bo
             <div className='dropdown_toggle' onClick={toggleDropdown}>
                 <img src={expanded ? UpIcon : DownIcon} className='icon medium pad' />
             </div>
+            {props.assignable &&
+                <div className='tag_container'>
+                    <div>Assigned tags:</div>
+                    {assignedTags && assignedTags.map((tag) => (
+                        <div className={`tag ${tag.color} ${selectedTag && selectedTag.id === tag.id ? 'active' : ''}`} key={tag.id} onClick={() => selectTag(tag)}>
+                            {tag.name}
+                        </div>
+                    ))}
+                </div>
+            }
             <div className='tag_container'>
+                <div>All tags:</div>
                 {tags && tags.map((tag) => (
                     <div className={`tag ${tag.color} ${selectedTag && selectedTag.id === tag.id ? 'active' : ''}`} key={tag.id} onClick={() => selectTag(tag)}>
                         {tag.name}
@@ -112,8 +150,13 @@ export const TagView = (props: { value: Tag[], productId: string, assignable: bo
                 </div>
                 <div className='buttons'>
                     {props.assignable &&
-                        <button className='button fill green' onClick={updateTag}>
+                        <button className='button fill green' onClick={assignTag}>
                             assign
+                        </button>
+                    }
+                    {props.assignable &&
+                        <button className='button fill red' onClick={deassignTag}>
+                            deassign
                         </button>
                     }
                     <button className='button fill green inline' onClick={addTag}>
